+++ app/agent/workflows/tests/workflows.ts
@@ -1,4 +1,4 @@
 export * from "~/agent/workflows";
 
-export { cancellationTestWorkflow } from "./wait-for-workflow/workflow";
+export { cancellationTestWorkflow } from "./cancellation-experiments/workflow";
 export { testLock, signalWorkflow, acquireLockAndWaitForSignal } from "./mutex/workflow";
